@@ -1,58 +1,58 @@
 {
-	"name": "ng-loaders",
-	"version": "3.1.0",
-	"scripts": {
-		"ng": "ng",
-		"start": "ng serve",
-		"build": "ng build",
-		"build:lib": "ng build --prod angular-loaders-lib",
-		"pack:lib": "cd dist/angular-loaders-lib && npm pack",
-		"package": "npm run build:lib && npm run pack:lib",
-		"deploy": "npm run package && npm publish dist/angular-loaders-lib",
-		"test": "ng test",
-		"lint": "ng lint",
-		"e2e": "ng e2e",
-		"postinstall": "ngcc"
-	},
-	"private": true,
-	"dependencies": {
-		"@angular/animations": "~9.1.1",
-		"@angular/common": "~9.1.1",
-		"@angular/compiler": "~9.1.1",
-		"@angular/core": "~9.1.1",
-		"@angular/forms": "~9.1.1",
-		"@angular/localize": "^9.1.13",
-		"@angular/platform-browser": "~9.1.1",
-		"@angular/platform-browser-dynamic": "~9.1.1",
-		"@angular/router": "~9.1.1",
-		"bootstrap": "^4.1.2",
-		"ng-loaders": "^2.0.0",
-		"rxjs": "~6.5.4",
-		"tslib": "^1.10.0",
-		"zone.js": "~0.10.2"
-	},
-	"devDependencies": {
-		"@angular-devkit/build-angular": "~0.901.1",
-		"@angular-devkit/build-ng-packagr": "~0.901.9",
-		"@angular/cli": "~9.1.1",
-		"@angular/compiler-cli": "~9.1.1",
-		"@angular/language-service": "~9.1.1",
-		"@types/jasmine": "~3.5.0",
-		"@types/jasminewd2": "~2.0.3",
-		"@types/node": "^12.11.1",
-		"codelyzer": "^5.1.2",
-		"jasmine-core": "~3.5.0",
-		"jasmine-spec-reporter": "~4.2.1",
-		"karma": "^5.1.0",
-		"karma-chrome-launcher": "~3.1.0",
-		"karma-coverage-istanbul-reporter": "~2.1.0",
-		"karma-jasmine": "~3.0.1",
-		"karma-jasmine-html-reporter": "^1.4.2",
-		"ng-packagr": "^9.0.0",
-		"protractor": "^7.0.0",
-		"ts-node": "~8.3.0",
-		"tsickle": "^0.38.1",
-		"tslint": "~6.1.0",
-		"typescript": "~3.8.3"
-	}
+  "name": "ng-loaders",
+  "version": "3.1.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "build:lib": "ng build --prod angular-loaders-lib",
+    "pack:lib": "cd dist/angular-loaders-lib && npm pack",
+    "package": "npm run build:lib && npm run pack:lib",
+    "deploy": "npm run package && npm publish dist/angular-loaders-lib",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "postinstall": "ngcc"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~10.2.4",
+    "@angular/common": "~10.2.4",
+    "@angular/compiler": "~10.2.4",
+    "@angular/core": "~10.2.4",
+    "@angular/forms": "~10.2.4",
+    "@angular/localize": "^10.2.4",
+    "@angular/platform-browser": "~10.2.4",
+    "@angular/platform-browser-dynamic": "~10.2.4",
+    "@angular/router": "~10.2.4",
+    "bootstrap": "^4.1.2",
+    "ng-loaders": "^2.0.0",
+    "rxjs": "~6.5.4",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.2"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.1002.3",
+    "@angular-devkit/build-ng-packagr": "~0.1002.3",
+    "@angular/cli": "~10.2.3",
+    "@angular/compiler-cli": "~10.2.4",
+    "@angular/language-service": "~10.2.4",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "ng-packagr": "^10.1.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tsickle": "^0.39.1",
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.7"
+  }
 }
